+++ Cargo.toml
@@ -39,6 +39,7 @@ members = [
     "provider/core/macros",
     "provider/datagen",
     "provider/fs",
+    "provider/registry",
 
     # Baked data
     "provider/baked/calendar",
@@ -151,6 +152,7 @@ icu_provider_macros = { version = "~1.5.0", path = "provider/core/macros", defau
 icu_provider_adapters = { version = "~1.5.0", path = "provider/adapters", default-features = false }
 icu_provider_blob = { version = "~1.5.0", path = "provider/blob", default-features = false }
 icu_provider_fs = { version = "~1.5.0", path = "provider/fs/", default-features = false }
+icu_registry = { version = "~1.5.0", path = "provider/registry/", default-features = false }
 
 # Baked data
 icu_calendar_data = { version = "~1.5.0", path = "provider/baked/calendar", default-features = false }
