+++ provider/datagen/src/baked_exporter.rs
@@ -802,4 +802,4 @@ macro_rules! cb {
         }
     }
 }
-crate::registry!(cb);
+icu_registry::registry!(cb);
