+++ tools/depcheck/src/allowlist.rs
@@ -162,6 +162,7 @@ pub const EXTRA_DATAGEN_DEPS: &[&str] = &[
     "icu_codepointtrie_builder",
     "icu_pattern",
     "icu_provider_adapters",
+    "icu_registry",
     "itertools",
     "itoa",
     "matrixmultiply",
